build_requirements.txt
GitPython==3.1.27
pyinstaller
